+++ Lmyc/Models/ErrorViewModel.cs
@@ -1,4 +1,5 @@
 using System;
+using System.ComponentModel.DataAnnotations;
 
 namespace Lmyc.Models
 {
@@ -7,5 +8,11 @@ namespace Lmyc.Models
         public string RequestId { get; set; }
 
         public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
+
+        [Display(Name = "Error")]
+        public string Error { get; set; }
+
+        [Display(Name = "Description")]
+        public string ErrorDescription { get; set; }
     }
 }