+++ Lmyc/Startup.cs
@@ -13,6 +13,7 @@ using Lmyc.Models;
 using Lmyc.Services;
 using Swashbuckle.AspNetCore.Swagger;
 using System.IO;
+using AspNet.Security.OpenIdConnect.Primitives;
 
 namespace Lmyc
 {
@@ -28,18 +29,28 @@ namespace Lmyc
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddDbContext<ApplicationDbContext>(options =>
-                options.UseSqlServer(Configuration.GetConnectionString("DefaultConnection"), b => b.MigrationsAssembly("Lmyc")));
+            services.AddMvc();
 
+            //services.AddDbContext<ApplicationDbContext>(options =>
+            //{
+            //    options.UseSqlServer(Configuration.GetConnectionString("DefaultConnection"), b => b.MigrationsAssembly("Lmyc"));
+
+            //    // Register the entity sets needed by OpenIddict.
+            //    // Note: use the generic overload if you need
+            //    // to replace the default OpenIddict entities.
+            //    options.UseOpenIddict();
+            //});
+          
             //var host = Configuration["DBHOST"] ?? "localhost";
             //var port = Configuration["DBPORT"] ?? "3306";
             //var password = Configuration["DBPASSWORD"] ?? "secret";
 
-            //services.AddDbContext<ApplicationDbContext>(options =>
-            //{
-            //    options.UseMySql($"server={host}; userid=root; pwd={password};"
-            //        + $"port={port}; database=lmyc");
-            //});
+            services.AddDbContext<ApplicationDbContext>(options =>
+            {
+                options.UseMySql($"server={host}; userid=root; pwd={password};"
+                    + $"port={port}; database=lmyc");
+                options.UseOpenIddict();
+            });
 
             services.AddIdentity<ApplicationUser, IdentityRole>()
                 .AddEntityFrameworkStores<ApplicationDbContext>()
@@ -48,7 +59,39 @@ namespace Lmyc
             // Add application services.
             services.AddTransient<IEmailSender, EmailSender>();
 
-            services.AddMvc();
+            // Register the OAuth2 validation handler.
+            services.AddAuthentication()
+                .AddOAuthValidation();
+
+            services.Configure<IdentityOptions>(options =>
+            {
+                options.ClaimsIdentity.UserNameClaimType = OpenIdConnectConstants.Claims.Name;
+                options.ClaimsIdentity.UserIdClaimType = OpenIdConnectConstants.Claims.Subject;
+                options.ClaimsIdentity.RoleClaimType = OpenIdConnectConstants.Claims.Role;
+            });
+
+            // Register the OpenIddict services.
+            // Note: use the generic overload if you need
+            // to replace the default OpenIddict entities.
+            services.AddOpenIddict(options =>
+            {
+                // Register the Entity Framework stores.
+                options.AddEntityFrameworkCoreStores<ApplicationDbContext>();
+
+                // Register the ASP.NET Core MVC binder used by OpenIddict.
+                // Note: if you don't call this method, you won't be able to
+                // bind OpenIdConnectRequest or OpenIdConnectResponse parameters.
+                options.AddMvcBinders();
+
+                // Enable the token endpoint (required to use the password flow).
+                options.EnableTokenEndpoint("/connect/token");
+
+                // Allow client applications to use the grant_type=password flow.
+                options.AllowPasswordFlow();
+
+                // During development, you can disable the HTTPS requirement.
+                options.DisableHttpsRequirement();
+            });
 
             // Register the Swagger generator, defining one or more Swagger documents
             services.AddSwaggerGen(c =>